Move variable products logic to add-to-cart-with-options/variation-selector script
