+++ plugins/woocommerce/client/blocks/assets/js/blocks/add-to-cart-with-options/variation-selector/attribute-options/block.json
@@ -26,6 +26,5 @@
 		"woocommerce/attributeName",
 		"woocommerce/attributeTerms"
 	],
-	"viewScriptModule": "woocommerce/add-to-cart-with-options-variation-selector-attribute-options",
 	"style": "file:../woocommerce/add-to-cart-with-options-variation-selector-attribute-options-style.css"
 }
